requirements.txt
streamlit==1.35.0
pandas
gcsfs
st-files-connection
Pillow==10.3.0
google.cloud
google.oauth2
